Refactor naming to ApplicationUiSettingGroup
