+++ applications/Unity.GrantManager/src/Unity.GrantManager.Web/Components/ApplicationUiSettingGroup/ApplicationUiSettingGroupViewComponent.cs
@@ -3,7 +3,7 @@ using System.Threading.Tasks;
 using Unity.GrantManager.SettingManagement;
 using Volo.Abp.AspNetCore.Mvc;
 
-namespace Unity.GrantManager.Web.Components.ApplicationTabsSettingGroup;
+namespace Unity.GrantManager.Web.Components.ApplicationUiSettingGroup;
 
 public class ApplicationUiSettingGroupViewComponent(IApplicationUiSettingsAppService settingsAppService) : AbpViewComponent
 {
@@ -11,6 +11,6 @@ public class ApplicationUiSettingGroupViewComponent(IApplicationUiSettingsAppSer
     {
         var model = await settingsAppService.GetAsync();
         var viewModel = ObjectMapper.Map<ApplicationUiSettingsDto, ApplicationUiSettingsViewModel>(model!);
-        return View("~/Components/ApplicationTabsSettingGroup/Default.cshtml", viewModel);
+        return View("~/Components/ApplicationUiSettingGroup/Default.cshtml", viewModel);
     }
 }
