+++ applications/Unity.GrantManager/src/Unity.GrantManager.Web/Components/ApplicationUiSettingGroup/Default.cshtml
@@ -1,6 +1,6 @@
-﻿@model Unity.GrantManager.Web.Components.ApplicationTabsSettingGroup.ApplicationUiSettingsViewModel
+﻿@model Unity.GrantManager.Web.Components.ApplicationUiSettingGroup.ApplicationUiSettingsViewModel
 
-<abp-script src="/Components/ApplicationTabsSettingGroup/Default.js" />
+<abp-script src="/Components/ApplicationUiSettingGroup/Default.js" />
 
 <div class="container-fluid px-0">
     <div class="unity-page-titlebar">
